+++ package.json
@@ -16,7 +16,9 @@
   ],
   "keywords": [
     "javascript",
-    "mock", "mocking", "mocks",
+    "mock",
+    "mocking",
+    "mocks",
     "simple",
     "strings",
     "generator",
